+++ src/extensions/gfmHeadingId/gfmHeadingId.extension.ts
@@ -1,4 +1,4 @@
-import {inject, Injectable, OnDestroy} from '@angular/core';
+import {Inject, inject, Injectable, OnDestroy, Optional} from '@angular/core';
 import {NavigationEnd, Router} from '@angular/router';
 import {MarkdownRendererExtension} from '@anglr/md-help';
 import {validHtmlId} from '@jscrpt/common';
@@ -40,7 +40,7 @@ export class GfmHeadingIdExtension implements MarkdownRendererExtension, OnDestr
     }
 
     //######################### constructor #########################
-    constructor()
+    constructor(@Inject({}) @Optional() withLinks?: boolean)
     {
         const route = inject(Router);
 
@@ -62,7 +62,10 @@ export class GfmHeadingIdExtension implements MarkdownRendererExtension, OnDestr
             {
                 heading({depth, text, tokens}: Tokens.Heading)
                 {
-                    return `<h${depth} id="${slugger.slug(validHtmlId(text))}">${this.parser.parseInline(tokens)}</h${depth}>`;
+                    const id = slugger.slug(validHtmlId(text));
+                    const link = `<a class="heading-anchor" href="${id}"><svg class="anchor-svg" viewBox="0 0 16 16" version="1.1" width="16" height="16" aria-hidden="true"><path d="m7.775 3.275 1.25-1.25a3.5 3.5 0 1 1 4.95 4.95l-2.5 2.5a3.5 3.5 0 0 1-4.95 0 .751.751 0 0 1 .018-1.042.751.751 0 0 1 1.042-.018 1.998 1.998 0 0 0 2.83 0l2.5-2.5a2.002 2.002 0 0 0-2.83-2.83l-1.25 1.25a.751.751 0 0 1-1.042-.018.751.751 0 0 1-.018-1.042Zm-4.69 9.64a1.998 1.998 0 0 0 2.83 0l1.25-1.25a.751.751 0 0 1 1.042.018.751.751 0 0 1 .018 1.042l-1.25 1.25a3.5 3.5 0 1 1-4.95-4.95l2.5-2.5a3.5 3.5 0 0 1 4.95 0 .751.751 0 0 1-.018 1.042.751.751 0 0 1-1.042.018 1.998 1.998 0 0 0-2.83 0l-2.5 2.5a1.998 1.998 0 0 0 0 2.83Z"></path></svg></a>`;
+
+                    return `<div class="relative"><h${depth} id="${id}">${this.parser.parseInline(tokens)}</h${depth}>${withLinks ? link : ''}</div>`;
                 }
             }
         };
@@ -76,4 +79,23 @@ export class GfmHeadingIdExtension implements MarkdownRendererExtension, OnDestr
     public ngOnDestroy(): void
     {
     }
+}
+
+/**
+ * Factory for GfmHeadingIdExtension with withLink parameter
+ * @param withLinks - Indication whether display link with headings
+ */
+export function gfmHeadingIdExtension(withLinks: boolean): typeof GfmHeadingIdExtension
+{
+    @Injectable()
+    class _GfmHeadingIdExtension extends GfmHeadingIdExtension
+    {
+        //######################### constructor #########################
+        constructor()
+        {
+            super(withLinks);
+        }
+    }
+
+    return _GfmHeadingIdExtension;
 }